Fix course information retrieval
If some of the course information could not be retrieved from the
EvaSys system, do not abort the whole retrieval process but instead
skip the respective course and continue with the other courses.

This is especially problematic for Moodle courses which are connected
with several EvaSys courses. Without this patch the block displays
"course missing" for all of the connected courses even if only one of
the courses does not exist in EvaSys.